pyproject.toml
[project]
name = "spiralmind-nexus"
version = "0.2.0"
description = "AI Double Pipeline System with Quantum-Inspired Processing"
authors = [
    {name = "Patryk Sobierański META-GENIUSZ®️🇮🇩", email = "patryksobieranski5@gmail.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
keywords = ["ai", "pipeline", "quantum", "processing", "spiral", "consciousness", "agi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

dependencies = [
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

test = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
spiral = "spiral.cli:main"

[project.urls]
Homepage = "https://github.com/sobieranskip95patryk/SpiralMind-Nexus"
Repository = "https://github.com/sobieranskip95patryk/SpiralMind-Nexus"
Issues = "https://github.com/sobieranskip95patryk/SpiralMind-Nexus/issues"
Documentation = "https://github.com/sobieranskip95patryk/SpiralMind-Nexus/blob/main/README.md"
Website = "https://www.mtaquestwebsidex.com"
Author = "https://github.com/sobieranskip95patryk"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["spiral*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false